🐛 readonly fields are not sent to next step
Field readonly was sent depending on having a value.
When the user gave a value, the field could not be
edited anymore.

Also readonly fields, are not used to update wizard
data so whichever we store as readonly is received
as False.

So we decoupled the real value of the fields from
the attribute telling that it should be readonly.
And considering that we could receive False for the values.

Co-authored-by: David García Garzón <david.garcia@somenergia.coop>